temp

+++ download.py
@@ -1,11 +1,12 @@
 """
 下载数据
-    jibenmian: 下载基本面数据，支持 资产负债表，利润表，现金流量表
+    jibenmian: 下载基本面数据（业绩快报），支持 资产负债表，利润表，现金流量表
     stock_daily: 下载股票日线
 """
 import akshare as ak
 import pandas as pd
 import time
+import os
 
 
 # ----- 下载所有股票基本面概要数据数据（业绩快报） ------
@@ -34,6 +35,12 @@ def jibenmian(type, yStart, yEnd, season=["{year}0331", "{year}0630", "{year}093
             time.sleep(3)
 
 
+def jibenmian_details():
+    stockIds = all_a_stocks()
+    for stockId in stockIds:
+        jibenmian_stock(stockId, 0)
+
+
 # ----- 下载 一只股票的基本面数据 ------
     """
     ak.stock_balance_sheet_by_report_em(symbol) # 资产负债表-按报告期
@@ -45,13 +52,26 @@ def jibenmian(type, yStart, yEnd, season=["{year}0331", "{year}0630", "{year}093
     ak.stock_cash_flow_sheet_by_report_em(symbol) # 现金流量表-按报告期
     ak.stock_cash_flow_sheet_by_yearly_em(symbol) # 现金流量表-年度
     """
-def jibenmian_stock(symbol):
+# symbol: 股票代码 格式: SHXXXXX or SZXXXXXX
+# delay: 每下载一个文件延迟的时间(秒)
+def jibenmian_stock(symbol, delay=3):
     funcMap = {'zcfz_report': ak.stock_balance_sheet_by_report_em, 
-               'zcfz_year': ak.stock_balance_sheet_by_yearly_em}
+               'lrb_report' : ak.stock_profit_sheet_by_report_em,
+               'xjll_report': ak.stock_cash_flow_sheet_by_report_em}
     for type, func in funcMap.items():
-        df = func(symbol)
         filename = f'data/stock/{type}_{symbol}.csv'
-        df.to_csv(filename)
+        if not os.path.exists(filename):
+            try:
+                df = func(symbol)
+                df.to_csv(filename)
+                if(delay > 0):
+                    time.sleep(delay)
+            except:
+                print(f"exec error: stock={symbol}")
+            else:
+                print(f"exec ok: stock={symbol}")
+        else:
+            print(f"skip: stock={symbol}, type={type}")
 
 # ------ 下载股票日线数据 ------
 #def stock_daily():
@@ -66,7 +86,7 @@ def all_a_stocks():
     stockIds = map(StockFormat, df['股票代码'])
     return stockIds
 
-# 数字转为 股票代码
+# 数字转为 股票代码 例如: 603339->'SH603339', 2985->'SZ002985'
 def StockFormat(stockId) :
     stockId = str(stockId).zfill(6)
     if stockId.startswith(('0', '3')):
@@ -78,4 +98,4 @@ def StockFormat(stockId) :
     return stockId
 
 # jibenmian('xjll', 2022, 2023)
-jibenmian_stock('SH600557')
+jibenmian_details()
